requirements.txt
selenium 3.141.0
pytest 6.2.1
pytest-html 3.1.1
allure-pytest 2.8.29
openpyxl 3.0.5
pytest-xdist 2.2.0